# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"AI Engineering Intelligence Platform","description":"A repository intelligence platform powered by static analysis, knowledge graphs, hybrid search, and LLM reasoning.","version":"0.1.0"},"paths":{"/repositories/":{"get":{"tags":["Repositories"],"summary":"List Repositories","operationId":"list_repositories_repositories__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/repositories/{repository_id}":{"delete":{"tags":["Repositories"],"summary":"Delete Repository","operationId":"delete_repository_repositories__repository_id__delete","parameters":[{"name":"repository_id","in":"path","required":true,"schema":{"type":"integer","title":"Repository Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/repositories/analyze":{"post":{"tags":["Repositories"],"summary":"Analyze Repository","operationId":"analyze_repository_repositories_analyze_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RepositoryAnalyzeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RepositoryAnalyzeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/graph/repositories/{repository_id}":{"get":{"tags":["Graph"],"summary":"Get Repository Graph","operationId":"get_repository_graph_graph_repositories__repository_id__get","parameters":[{"name":"repository_id","in":"path","required":true,"schema":{"type":"integer","title":"Repository Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/graph/files/{file_id}/symbols":{"get":{"tags":["Graph"],"summary":"Get File Symbols","operationId":"get_file_symbols_graph_files__file_id__symbols_get","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"integer","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/graph/symbols/{symbol_id}/callers":{"get":{"tags":["Graph"],"summary":"Get Symbol Callers","operationId":"get_symbol_callers_graph_symbols__symbol_id__callers_get","parameters":[{"name":"symbol_id","in":"path","required":true,"schema":{"type":"integer","title":"Symbol Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/graph/symbols/{symbol_id}/callees":{"get":{"tags":["Graph"],"summary":"Get Symbol Callees","operationId":"get_symbol_callees_graph_symbols__symbol_id__callees_get","parameters":[{"name":"symbol_id","in":"path","required":true,"schema":{"type":"integer","title":"Symbol Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/graph/files/{file_id}/dependencies":{"get":{"tags":["Graph"],"summary":"Get File Dependencies","operationId":"get_file_dependencies_graph_files__file_id__dependencies_get","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"integer","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/graph/symbols/{symbol_id}/parents":{"get":{"tags":["Graph"],"summary":"Get Parent Classes","operationId":"get_parent_classes_graph_symbols__symbol_id__parents_get","parameters":[{"name":"symbol_id","in":"path","required":true,"schema":{"type":"integer","title":"Symbol Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/graph/symbols/{symbol_id}/children":{"get":{"tags":["Graph"],"summary":"Get Child Classes","operationId":"get_child_classes_graph_symbols__symbol_id__children_get","parameters":[{"name":"symbol_id","in":"path","required":true,"schema":{"type":"integer","title":"Symbol Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/embeddings/generate":{"post":{"tags":["Embeddings"],"summary":"Generate Embeddings","operationId":"generate_embeddings_embeddings_generate_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/embeddings/search":{"get":{"tags":["Embeddings"],"summary":"Search Code","operationId":"search_code_embeddings_search_get","parameters":[{"name":"repository_id","in":"query","required":true,"schema":{"type":"integer","title":"Repository Id"}},{"name":"q","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Q"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reasoning/ask":{"post":{"tags":["Reasoning"],"summary":"Ask Codebase","operationId":"ask_codebase_reasoning_ask_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReasoningRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ReasoningRequest":{"properties":{"query":{"type":"string","title":"Query"},"repository_id":{"type":"integer","title":"Repository Id"},"limit":{"type":"integer","title":"Limit","default":5}},"type":"object","required":["query","repository_id"],"title":"ReasoningRequest"},"RepositoryAnalyzeRequest":{"properties":{"repository_url":{"type":"string","maxLength":2083,"minLength":1,"format":"uri","title":"Repository Url"}},"type":"object","required":["repository_url"],"title":"RepositoryAnalyzeRequest"},"RepositoryAnalyzeResponse":{"properties":{"repository_id":{"type":"integer","title":"Repository Id"},"github_url":{"type":"string","title":"Github Url"},"local_path":{"type":"string","title":"Local Path"},"status":{"type":"string","title":"Status"}},"type":"object","required":["repository_id","github_url","local_path","status"],"title":"RepositoryAnalyzeResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}